5 - Search and Replace
First argument is a sentence to perform a search and replace.  Second
argument is the word that will be replaced with the third argument.  The
case of the word that is being replaced must be preserved.  Function
determines if the first letter in the second argument is uppercase, if
true, it will change the first letter in the third argument to
uppercase.  It will then replace the changed or unchanged third argument
with the second argument in the sentence.