6 - Pig Latin
Translates a string to pig latin.  Creates an array of vowels.  Function
first determines if the string begins with a vowel.  If true, returns
the original string with "way" added to the end.  If false, using a
while and for loop, it iterates through the vowel array and the string
to determine the index of the first vowel.  It then creates and returns
a new string that includes the first vowel and remainder of the word +
the consonants before the first vowel + 'ay'.