8 - Missing Letters
Function takes in a string of letters, returning the missing letter or,
if there is no missing letter, undefined.  The function first converts
the string to an array of unicode.  It then iterates through the array,
comparing the current array number+1 to the next array number.  Once
they are not equal, it has found the missing unicode number of the
letter, which it then converts from unicode to a string containing the
missing. letter.  If they are never found not to be equal, the function
returns undefined.