15 - Smallest Common Multiple - Flagged to revisit
Function finds the smallest common multiple of two provided parameters
that can be evenly divided by both, as well as by all sequential numbers
in the range between these parameters.  The code is horrible
inefficient.  In fact, if you run it in freeCodeCamp it will stop with
an infinite loop warning - you have to add //noprotect on the first line
for it to work.  There isn't an infinite loop, it just takes forever to
process.  I  want to redo this one to take advantage of arrays, could
make this a lot better.   Here, the for-i loop finds the LCD.  The for-j
loop then determines if the LCD is divisible by all the numbers in the
parameter range.  If it is, all the loops exit, we're done.  If not, the
for-i loop continues in finding the next common denominator, then
executes the j-for loop again, repeat until success.