Symmetric Difference
Easily a better way to do this.  I didn't understand the task as good as
I thought I did when I started, so there are a bunch of band-aids to fix
issues I had and variable names that don't make much sense.   The
function takes two or more arrays and returns an array of the symmetric
difference of the provided ways.   It joins both arrays together and
comes up with a list of duplicate numbers to remove.  It then determines
if the duplicates are present in both arguments.  If so, it removes
these numbers.  It then removes any duplicates that are left over (as
they were duplicates in a single argument).