Exact Change
Function accepts a purchase price as the first argument, payment as the
second, and a cash-in-drawer as the third.  Function returns
"Insufficient Funds" if change can't be made using available
cash-in-drawer.  It returns "closed" if cash-in-drawer is equal to
change due.  Otherwise, it returns an array of exact change in coins and
bills, sorted in highest to lowest order.