Inventory Update
An inventory array is passed as the first argument and an update to it
as the second argument.  If the item in the update already exists in the
inventory, the quantity of the item is updated in the original
inventory.  If the item does not exist, the item name and quantity are
added to the original inventory in alphabetical order.