Friendly Date Ranges
Functions takes in an array of 2 dates in YYYY-MM-DD.  Returns a
"friendly range" with month numbers converted to month names and
cardinal dates converted to ordinal dates.  "Friendly range" means: If
date range ends in less than a year from when it begins, do not display
ending year.  If the range ends in the same month that it begins, do no
display the ending year or month.  If the date range begins in the
current year and ends within a year, the year is not displayed at the
beginning of the range.