Pairwise
Returns the sum of all element indices of the first argument array that
can be paired with one other element to form a sum that equals the value
in the second argument.  If multiple sums are possible, it returns the
smallest sum.  Once an element has been used, it cannot be reused to
parit with another.